requirements.txt
amqp==5.2.0
asgiref==3.8.1
async-timeout==4.0.3
billiard==4.2.0
boto3==1.34.139
botocore==1.34.139
celery==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==5.0.6
django-admin-adminlte==1.0.5
django-anymail==11.0.1
django-appconf==1.0.6
django-celery-email==3.0.0
django-ckeditor-5==0.2.13
django-js-asset==2.2.0
django-jsoneditor==0.2.4
django-modeltranslation==0.19.7
django-multiselectfield==0.1.13
django-phonenumber-field==7.3.0
django-storages==1.14.3
Faker==27.0.0
idna==3.7
jmespath==1.0.1
kombu==5.3.7
mysqlclient==2.2.4
olefile==0.47
packaging==24.1
phonenumbers==8.13.39
pillow==10.3.0
prompt_toolkit==3.0.47
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
redis==5.0.8
requests==2.32.3
s3transfer==0.10.2
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
